requirements.txt
# 個人訊息流強化系統 - Python 依賴

# API 客戶端
requests>=2.28.0
anthropic>=0.18.0

# 環境變數管理
python-dotenv>=1.0.0

# Readwise API (可選，目前使用直接 requests)
# readwise-api>=0.3.0
